+++ vnext/Shared/OInstance.cpp
@@ -28,6 +28,7 @@
 #include <Modules/ExceptionsManagerModule.h>
 #include <Modules/PlatformConstantsModule.h>
 #include <Modules/SourceCodeModule.h>
+#include <Modules/StatusBarManagerModule.h>
 
 #if (defined(_MSC_VER) && (defined(WINRT)))
 #include <Utils/LocalBundleReader.h>
@@ -649,12 +650,18 @@ std::vector<std::unique_ptr<NativeModule>> InstanceImpl::GetDefaultNativeModules
       []() { return std::make_unique<PlatformConstantsModule>(); },
       nativeQueue));
 
+  modules.push_back(std::make_unique<CxxNativeModule>(
+      m_innerInstance,
+      StatusBarManagerModule::Name,
+      []() { return std::make_unique<StatusBarManagerModule>(); },
+      nativeQueue));
+
   return modules;
 }
 
 void InstanceImpl::RegisterForReloadIfNecessary() noexcept {
   // setup polling for live reload
-  if (!m_devManager->HasException() && m_devSettings->liveReloadCallback != nullptr) {
+  if (!m_devManager->HasException() && !m_devSettings->useFastRefresh && m_devSettings->liveReloadCallback != nullptr) {
     m_devManager->StartPollingLiveReload(m_devSettings->debugHost, m_devSettings->liveReloadCallback);
   }
 }